# backend/docs/swagger/swagger.json
{
    "basePath": "/api/v1",
    "definitions": {
...
        "ErrorResponse": {
            "properties": {
                "code": {
                    "example": "TEAM_NOT_FOUND",
                    "type": "string"
                },
                "error": {
                    "example": "Team not found",
                    "type": "string"
                },
                "message": {
                    "example": "No team exists with id team-unknown",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "LoginResponse": {
            "properties": {
                "accessToken": {
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJkZW1vIn0.signature",
                    "type": "string"
                },
                "expiresIn": {
                    "description": "Access token expiry in seconds",
                    "example": 900,
                    "type": "integer"
                },
                "refreshToken": {
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlblR5cGUiOiJyZWZyZXNoIn0.signature",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/UserDTO"
                }
            },
            "type": "object"
        },
...
        "SSOCallbackRequest": {
            "properties": {
                "code": {
                    "example": "4/0AY0e-g6i9XqZ8...",
                    "type": "string"
                },
                "code_verifier": {
                    "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk",
                    "type": "string"
                }
            },
            "required": [
                "code",
                "code_verifier"
            ],
            "type": "object"
        },
...
        "UserDTO": {
            "properties": {
                "canTakeSurvey": {
                    "example": true,
                    "type": "boolean"
                },
                "email": {
                    "example": "demo@teams360.demo",
                    "type": "string"
                },
                "fullName": {
                    "example": "Demo User",
                    "type": "string"
                },
                "hierarchyLevel": {
                    "example": "level-5",
                    "type": "string"
                },
                "id": {
                    "example": "demo",
                    "type": "string"
                },
                "teamIds": {
                    "example": [
                        "team-phoenix",
                        "team-titan"
                    ],
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "username": {
                    "example": "demo",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "version": "1.0"
    },
    "paths": {
...
        "/auth/sso/callback": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Completes the OAuth PKCE flow: exchanges the authorization code and code verifier for the provider's tokens, reads the email claim from the returned ID token, and looks up a matching local user by that email. If the account exists and is configured for SSO, issues our own JWT access/refresh token pair in the same shape as a password login. Fails with a 503 if SSO is not configured on the server, and with a 401 if the token exchange fails, no account matches the email, or the matched account is not an SSO account. This is a public endpoint that does not require a prior session.",
                "parameters": [
                    {
                        "description": "Authorization code and PKCE verifier",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SSOCallbackRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Missing code or code_verifier",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token exchange failed, no account found, or account does not support SSO",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate authentication tokens",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "SSO is not configured on this server",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "summary": "SSO OAuth callback",
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}